fabtests/shared.c: Changed Memory Registration for larger buffers.
For the RMA multinode tests, to avoid having to register memory for
each process to do sends and receives in, one large buffer is
registered instead. The test then divides up that buffer and assigns
each process a portion of it to do sends and receives in. This helps
avoid having multiple processes writing and reading in the same spot.

Signed-off-by: Alex McKinley <alex@mckpals.com>
Signed-off-by: Alex McKinley <alex.mckinley@intel.com>