fabtests/common: fix atomic buffer
ft_post_atomic posted "buf" which is the base address for the entire
send and recv buffer allocation. The first half of the allocation is the
receive buffer and the second half is the send buffer. Posting just "buf"
meant it was sending the receive buffer.
This changes it to send the tx buf and do an atomic on the rx buf which allows
us to properly do atomic validation

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>