prov/shm: Add HMEM fallback protocol to enable CMA with HMEM
CMA: "Cross Memory Attach"
HMEM: "Heterogeneous Memory"

CMA was not available before these changes because a sender
could not guarentee that the receiving buffer is the same
type of memory as the sending one.

This protocol now allows the sender to send a
Host (System) Memory buffer with HMEM enabled.
This is because in the case where the reciever realizes
the destination is a Device Memory buffer, the reciever
will create the necessary resources for doing a
H->D (Host to Device) copy and return the command to the
sender's return queue. The sender will then do the
H->D copy and resubmit the command as an
ACK (Acknowledgement) operation. When the reciever gets
this ACK command it will cleanup the all reciever resources.
This includes the ones that it opened in order for the
sender to do the copy. It will pass the command back one
more time to the sender's return queue to complete the
transmission and cleanup the sender's resources.

This protocol also now allows for multiple iovs to be sent
with HMEM enabled. Example case: iov[0] is system memory,
iov[1] is ze memory, and iov[2] is cuda memory.
In this case the command will be re-used and sent back and
forth until all the copies are completed. They will take
the necessary fallback paths and set the "command ack" as
needed. The protocol will only complete the transmission
when the bytes_done equals the expected amount of data
from the original command.

With this fallback protocol H->H copies can take the CMA
fastpath when using shm on a node with HMEM devices.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>