prov/shm: new shm
Turn response queue into return queue for local commands
Inline commands are still receive side
All commands have an inline option but a common ptr to the command
being used for remote commands. These commands have to be returned
to the sender but the receive side can hold onto them as long as
needed for the lifetime of the message

shm has self and peer caps for each p2p interface (right now just
CMA and xpmem). The support for each of these interfaces is saved
in separate fields which causes a lot of wasted memory and is
confusing. Merge these into two fields (one for self and one for
peer) which holds the information for all p2p interfaces and is
accessed by the P2P type enums. CMA also needs a flag to know wether
CMA support has been queried yet or not.

This also moves some shm fields around for alignment

Simplifies access to the map to remove need for container

There is a 1:1 relationship with the av and map so just reuse
the util av lock for access to the map as well.
This requires some reorganizing of the locking semantics

There is nothing in smr_fabric, just use the util_fabric directly

Just like on the send side, make the progress functions be an array
of function pointers accessible by the command proto.
This cleans up the parameters of the progress calls and streamlines the
calls

This also renames the proto_ops to send_ops to make the two more clear

merge tx and pend entries for simple management of pending operations

proto data isn't really needed since it's only used for the inject offset
now and the cmd stack and inject buffers run in parallel.
Use a simple inject buf array and access by index where the index is the same
as the command's index in its stack

DSA copies happen asynchronously so we need a way to notify the receiver
when the copy is done and the data is available. This used to be done with
the response queue and sar list. The response queue notification can still be
done with the return of the command but the sar list was removed since more
data is sent by returning the command on a subsequent loop.
DSA still needs this list to track asynchronous copies. This refactors the
async ipc list and turns it into a generic async list to track asynchronous copies.
If a DSA copy is not ready, the entry is inserted into the async list and polled
until it is ready to be copied and then it resumes the regular SAR protocol where
the command is returned to the sender.

Tracking the status of the sar is done through the existing sar status of the peer
but to check for the correct status, the rx id is also needed by the receiver for
proper status exchange.

This also refactors the ids sto make it clearer when an id is used for the trasmitter (tx)
or target (rx)