prov/shm, include: Change inject resources from sender to receiver side
Add support to ofi_atomic_queue to find an entry pointer based on a buffer
from that entry. This is necessary so that the sender side of the inject
protocol can grab a receiver side resource to use for copying data to.
This allows the sender to complete when they are done and the receiver
can handle that acutal moving of the data.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>