prov/lnx: set core hints proto to UNSPEC
lnx has to overwrite some of the hints passed into the main getinfo call
because they won't be valid for all core providers (for example the
remote comm/local comm specification). This also includes the fabric and
domain names which won't match the core providers.

The protocol setting was missing so if the provider passed in FI_PROTO_LNX,
it would fail the core getinfos because of a mismatch. This resets the protocol
to a value which won't fail any core providers (FI_PROTO_UNSPEC)